backend/requirements.txt
Django==3.2.16
psycopg2-binary==2.9.3
python-dotenv==1.0.0
django-colorfield==0.7.2
djangorestframework==3.12.4
djoser==2.1.0
django-filter==23.5
gunicorn==20.1.0
tqdm==4.66
drf_extra_fields==3.3.0